Refactor seedHelpers.js and seed.js
